+++ src/Management/Application/StoringAndUpdatingFields.php
@@ -4,6 +4,7 @@
 namespace Thinktomorrow\Chief\Management\Application;
 
 use Illuminate\Http\Request;
+use Thinktomorrow\Chief\Fields\Types\Field;
 use Thinktomorrow\Chief\Fields\Types\FieldType;
 use Thinktomorrow\Chief\Fields\Fields;
 use Thinktomorrow\Chief\Management\Manager;
@@ -11,18 +12,15 @@ use Thinktomorrow\Chief\Management\Manager;
 trait StoringAndUpdatingFields
 {
     // If there is a save<key>Field this has priority over the set<Key>Field methods
+    protected $saveAssistantMethods = [];
     protected $saveMethods = [];
 
     protected function handleFields(Manager $manager, Request $request)
     {
-        foreach ($manager->fields() as $field) {
+        foreach ($manager->fieldsWithAssistantFields() as $field) {
 
             // Custom save methods
-            $saveMethodName = 'save'. ucfirst(camel_case($field->key())) . 'Field';
-            if (method_exists($manager, $saveMethodName)) {
-                $this->saveMethods[$field->key] = ['field' => $field, 'method' => $saveMethodName];
-                continue;
-            }
+            if($this->detectCustomSaveMethods($manager, $field)) continue;
 
             // Media fields are treated separately
             if ($field->ofType(FieldType::MEDIA, FieldType::DOCUMENT)) {
@@ -46,11 +44,37 @@ trait StoringAndUpdatingFields
         $manager->saveFields();
     }
 
+    protected function detectCustomSaveMethods(Manager $manager, Field $field): bool
+    {
+        $saveMethodName = 'save'. ucfirst(camel_case($field->key())) . 'Field';
+
+        // Custom save method on assistant
+        foreach($manager->assistants() as $assistant)
+        {
+            if (method_exists($assistant, $saveMethodName)) {
+                $this->saveAssistantMethods[$field->key] = ['field' => $field, 'method' => $saveMethodName, 'assistant' => $assistant];
+                return true;
+            }
+        }
+
+        // Custom save method on manager class
+        if (method_exists($manager, $saveMethodName)) {
+            $this->saveMethods[$field->key] = ['field' => $field, 'method' => $saveMethodName];
+            return true;
+        }
+
+        return false;
+    }
+
     protected function handleCustomSaves($manager, $request)
     {
+        foreach ($this->saveAssistantMethods as $data) {
+            $method = $data['method'];
+            $data['assistant']->$method($data['field'], $request);
+        }
+
         foreach ($this->saveMethods as $data) {
             $method = $data['method'];
-
             $manager->$method($data['field'], $request);
         }
     }